django-backend/requirements.txt
reaioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
constantly==15.1.0
cryptography==36.0.1
csp==0.1a10
daphne==3.0.2
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.11.0
django-csp==3.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.6.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.3
msgpack-python==0.5.6
nest-asyncio==1.5.4
platformdirs==2.5.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-dotenv==0.19.2
pytz==2021.3
redis==2.10.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
virtualenv==20.13.3
zope.interface==5.4.0